+++ CHANGELOG.md
@@ -127,3 +127,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Email Sender: ALL_DEVELOP, ALL_PILOT, ALL_FINAL
 - Send email to all involved users after a publication has been uploaded
 - Project and project view url in email context
+- Create notification after sending an email